자료형 수정

+++ FootPrints/src/main/java/footprints/footprints/controller/chat/ChatController.java
@@ -30,7 +30,7 @@ public class ChatController {
     }
 
     @GetMapping(value = "/chat/get-PostInfoList") // 각 방들의 post_name과 category post_id를 리스트(String) 형태로 가져온다.
-    public ResponseEntity<List<String>> getPostInfoList(@RequestBody List<Integer> postIdList){
+    public ResponseEntity<List<String>> getPostInfoList(@RequestBody List<Long> postIdList){
 
         List<String> postInfoList = chatService.getPostInfoList(postIdList);
 
@@ -38,7 +38,7 @@ public class ChatController {
     }
 
     @GetMapping(value = "/chat/get-NickList") // 그 방에 속한 사용자들의 nick을 리스트(String) 형태로 가져온다.
-    public ResponseEntity<List<String>> getNickList(@RequestBody int post_id){
+    public ResponseEntity<List<String>> getNickList(@RequestBody Long post_id){
 
         List<String> nickList = chatService.getNickList(post_id);
 
@@ -46,7 +46,7 @@ public class ChatController {
     }
 
     @GetMapping(value = "/chat/get-ChatList") // 채팅 기록 (from_name, time, message)를 리스트(String) 형태로 가져온다.
-    public ResponseEntity<List<String>> getChatList(@RequestBody int post_id){
+    public ResponseEntity<List<String>> getChatList(@RequestBody Long post_id){
 
         List<String> chatList = chatService.getChatList(post_id);
